Altar Global Progression System (pt.1)

+++ Assets/Project/Scripts/Configs/Buildings/BaseConstantsConfig.cs
@@ -5,8 +5,8 @@ namespace Configs
     public abstract class BaseConstantsConfig
     {
         public readonly string ID;
-        public readonly int value;
-        public BaseConstantsConfig(string id, int value)
+        public readonly float value;
+        public BaseConstantsConfig(string id, float value)
         {
             ID = id;
             this.value = value;
